Document `externalTypes` (#6962)
