+++ packages/server/src/externalTypes/graphql.ts
@@ -1,16 +1,12 @@
-import type { Trace } from '@apollo/usage-reporting-protobuf';
-import type {
-  DocumentNode,
-  FormattedExecutionResult,
-  GraphQLError,
-  GraphQLSchema,
-  OperationDefinitionNode,
-} from 'graphql';
-import type { CachePolicy } from '@apollo/cache-control-types';
+/**
+ * This file exports types related specifically to GraphQL execution, most
+ * notably `GraphQLRequest` and `GraphQLResponse`, used most heavily by
+ * `executeHTTPGraphQLRequest` and `executeOperation`. The
+ * `responseForOperation` plugin hook also returns a `GraphQLResponse`.
+ */
+import type { FormattedExecutionResult } from 'graphql';
 import type { BaseContext } from './context.js';
 import type { HTTPGraphQLHead, HTTPGraphQLRequest } from './http.js';
-import type { Logger } from '@apollo/utils.logger';
-import type { KeyValueCache } from '@apollo/utils.keyvaluecache';
 import type { WithRequired } from '@apollo/utils.withrequired';
 import type {
   GraphQLExperimentalFormattedInitialIncrementalExecutionResult,
@@ -54,59 +50,6 @@ export type GraphQLResponse<TData = Record<string, unknown>> = WithRequired<
   GraphQLInProgressResponse<TData>,
   'body'
 >;
-
-export interface GraphQLRequestMetrics {
-  // It would be more accurate to call this fieldLevelInstrumentation (it is
-  // true if the hook of that name returns truthy) but for backwards
-  // compatibility we keep this name (this field is read by `@apollo/gateway` to
-  // decide whether or not to send the apollo-federation-include-trace header).
-  captureTraces?: boolean;
-  persistedQueryHit?: boolean;
-  persistedQueryRegister?: boolean;
-  responseCacheHit?: boolean;
-  forbiddenOperation?: boolean;
-  registeredOperation?: boolean;
-  startHrTime?: [number, number];
-  queryPlanTrace?: Trace.QueryPlanNode;
-}
-
-export interface GraphQLRequestContext<TContext extends BaseContext> {
-  readonly logger: Logger;
-  readonly cache: KeyValueCache<string>;
-
-  readonly request: GraphQLRequest;
-  readonly response: GraphQLInProgressResponse;
-
-  readonly schema: GraphQLSchema;
-
-  readonly contextValue: TContext;
-
-  readonly queryHash?: string;
-
-  readonly document?: DocumentNode;
-  readonly source?: string;
-
-  // `operationName` is set based on the operation AST, so it is defined even if
-  // no `request.operationName` was passed in.  It will be set to `null` for an
-  // anonymous operation, or if `requestName.operationName` was passed in but
-  // doesn't resolve to an operation in the document.
-  readonly operationName?: string | null;
-  readonly operation?: OperationDefinitionNode;
-
-  /**
-   * Unformatted errors which have occurred during the request. Note that these
-   * are present earlier in the request pipeline and differ from **formatted**
-   * errors which are the result of running the user-configurable `formatError`
-   * transformation function over specific errors; these can eventually be found
-   * in `response.result.errors`.
-   */
-  readonly errors?: ReadonlyArray<GraphQLError>;
-
-  readonly metrics: GraphQLRequestMetrics;
-
-  readonly overallCachePolicy: CachePolicy;
-}
-
 export interface ExecuteOperationOptions<TContext extends BaseContext> {
   contextValue?: TContext;
 }
