+++ packages/server/src/externalTypes/index.ts
@@ -5,12 +5,7 @@
  * intention (it's public API).
  */
 export type { BaseContext, ContextFunction, ContextThunk } from './context.js';
-export type {
-  GraphQLRequest,
-  GraphQLRequestContext,
-  GraphQLRequestMetrics,
-  GraphQLResponse,
-} from './graphql.js';
+export type { GraphQLRequest, GraphQLResponse } from './graphql.js';
 export type {
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
@@ -31,6 +26,8 @@ export type {
   LandingPage,
 } from './plugins.js';
 export type {
+  GraphQLRequestContext,
+  GraphQLRequestMetrics,
   GraphQLRequestContextDidEncounterErrors,
   GraphQLRequestContextDidResolveOperation,
   GraphQLRequestContextDidResolveSource,
